requirements.txt
django==4.0.6
psycopg2==2.9.3
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap4==22.1